requirements.txt
astroid==2.3.3
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-phone-field==1.8.0
django-sendsms==0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.12
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
twilio==6.33.1
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
